Give files the correct names
